refactor: move functions where they belong and update tests accordingly
